Import latest version 4.4.
git-svn-id: https://svn.dealii.org/trunk@10997 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/contrib/umfpack/AMD/Source/amd_2.c
@@ -584,7 +584,7 @@ GLOBAL void AMD_2
     }
     else
     {
-	dense = (Int) (alpha * sqrt ((double) n)) ;
+	dense = alpha * sqrt ((double) n) ;
     }
     dense = MAX (16, dense) ;
     dense = MIN (n,  dense) ;
